+++ src/AutoMapper/Execution/TypeMapPlanBuilder.cs
@@ -1,0 +1,464 @@
+﻿namespace AutoMapper.Execution
+{
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Linq.Expressions;
+    using static System.Linq.Expressions.Expression;
+    using static ExpressionExtensions;
+    using System.Reflection;
+    using Configuration;
+    using Mappers;
+
+    public static class TypeMapPlanBuilder
+    {
+        public static LambdaExpression BuildMapperFunc(TypeMap typeMap, TypeMapRegistry typeMapRegistry)
+        {
+            if (typeMap.SourceType.IsGenericTypeDefinition() || typeMap.DestinationType.IsGenericTypeDefinition())
+                return null;
+
+            var srcParam = Parameter(typeMap.SourceType, "src");
+            var destParam = Parameter(typeMap.DestinationType, "dest");
+            var ctxtParam = Parameter(typeof(ResolutionContext), "ctxt");
+
+            if (typeMap.Substitution != null)
+            {
+                return Lambda(typeMap.Substitution.ReplaceParameters(srcParam, destParam, ctxtParam), srcParam, destParam, ctxtParam);
+            }
+
+            if (typeMap.TypeConverterType != null)
+            {
+                Type type;
+                if (typeMap.TypeConverterType.IsGenericTypeDefinition())
+                {
+                    var genericTypeParam = typeMap.SourceType.IsGenericType()
+                        ? typeMap.SourceType.GetTypeInfo().GenericTypeArguments[0]
+                        : typeMap.DestinationType.GetTypeInfo().GenericTypeArguments[0];
+                    type = typeMap.TypeConverterType.MakeGenericType(genericTypeParam);
+                }
+                else type = typeMap.TypeConverterType;
+
+                // (src, dest, ctxt) => ((ITypeConverter<TSource, TDest>)ctxt.Options.CreateInstance<TypeConverterType>()).Convert(src, ctxt);
+                Type converterInterfaceType = typeof(ITypeConverter<,>).MakeGenericType(typeMap.SourceType, typeMap.DestinationType);
+                return Lambda(
+                    Call(
+                        Convert(
+                            Call(
+                                MakeMemberAccess(ctxtParam, typeof(ResolutionContext).GetProperty("Options")),
+                                typeof(MappingOperationOptions).GetMethod("CreateInstance")
+                                    .MakeGenericMethod(type)
+                                ),
+                            converterInterfaceType),
+                        converterInterfaceType.GetMethod("Convert"),
+                        srcParam, ctxtParam
+                        ),
+                    srcParam, destParam, ctxtParam);
+
+            }
+
+            if (typeMap.CustomMapper != null)
+            {
+                return Lambda(typeMap.CustomMapper.ReplaceParameters(srcParam, destParam, ctxtParam), srcParam, destParam, ctxtParam);
+            }
+
+            if (typeMap.CustomProjection != null)
+            {
+                return Lambda(typeMap.CustomProjection.ReplaceParameters(srcParam), srcParam, destParam, ctxtParam);
+            }
+
+            var destinationFunc = CreateDestinationFunc(typeMap, typeMapRegistry, srcParam, destParam, ctxtParam);
+
+            var assignmentFunc = CreateAssignmentFunc(typeMap, typeMapRegistry, srcParam, destParam, ctxtParam, destinationFunc);
+
+            var mapperFunc = CreateMapperFunc(typeMap, srcParam, destParam, ctxtParam, assignmentFunc);
+
+            var lambdaExpr = Lambda(mapperFunc, srcParam, destParam, ctxtParam);
+
+            return lambdaExpr;
+        }
+
+        private static Expression CreateMapperFunc(
+            TypeMap typeMap,
+            ParameterExpression srcParam,
+            ParameterExpression destParam,
+            ParameterExpression ctxtParam,
+            Expression assignmentFunc)
+        {
+            var mapperFunc = assignmentFunc;
+
+            if (typeMap.Condition != null)
+            {
+                mapperFunc =
+                    Condition(Invoke(typeMap.Condition, ctxtParam),
+                        mapperFunc, Default(typeMap.DestinationType));
+                //mapperFunc = (source, context, destFunc) => _condition(context) ? inner(source, context, destFunc) : default(TDestination);
+            }
+
+            if (typeMap.Profile.AllowNullDestinationValues)
+            {
+                mapperFunc =
+                    Condition(Equal(srcParam, Default(typeMap.SourceType)),
+                        Default(typeMap.DestinationType), mapperFunc);
+                //mapperFunc = (source, context, destFunc) => source == default(TSource) ? default(TDestination) : inner(source, context, destFunc);
+            }
+
+            if (typeMap.PreserveReferences)
+            {
+                var cache = Variable(typeMap.DestinationType, "cachedDestination");
+
+                var condition = Condition(
+                    AndAlso(
+                        NotEqual(srcParam, Constant(null)),
+                        AndAlso(
+                        Equal(destParam, Constant(null)),
+                        Call(Property(ctxtParam, "InstanceCache"), typeof(Dictionary<object, object>).GetMethod("ContainsKey"), srcParam)
+                    )),
+                    Assign(cache, ToType(Property(Property(ctxtParam, "InstanceCache"), "Item", srcParam), typeMap.DestinationType)),
+                    Assign(cache, mapperFunc)
+                );
+
+                mapperFunc = Block(new[] { cache }, condition, cache);
+            }
+            return mapperFunc;
+        }
+
+        private static Expression CreateAssignmentFunc(
+            TypeMap typeMap,
+            TypeMapRegistry registry,
+            ParameterExpression srcParam,
+            ParameterExpression destParam,
+            ParameterExpression ctxtParam,
+            Expression destinationFunc)
+        {
+            var assignTypeMap = Assign(MakeMemberAccess(ctxtParam, typeof(ResolutionContext).GetProperty("TypeMap")), Constant(typeMap));
+
+            var beforeMap = Call(ctxtParam, typeof(ResolutionContext).GetMethod("BeforeMap"), ToObject(destParam));
+
+            var typeMaps =
+                typeMap.GetPropertyMaps().Where(pm => pm.CanResolveValue()).Select(pm => TryPropertyMap(pm, registry, srcParam, destParam, ctxtParam)).ToList();
+
+            var afterMap = Call(ctxtParam, typeof(ResolutionContext).GetMethod("AfterMap"), ToObject(destParam));
+
+            var actions = typeMaps;
+
+            foreach (var beforeMapAction in typeMap.BeforeMapActions)
+            {
+                actions.Insert(0, beforeMapAction.ReplaceParameters(srcParam, destParam, ctxtParam));
+            }
+            actions.Insert(0, beforeMap);
+            actions.Insert(0, destinationFunc);
+            actions.Insert(0, assignTypeMap);
+
+            actions.Add(afterMap);
+
+            actions.AddRange(typeMap.AfterMapActions.Select(afterMapAction => afterMapAction.ReplaceParameters(srcParam, destParam, ctxtParam)));
+
+            actions.Add(destParam);
+
+            return Block(actions);
+        }
+
+        private static Expression TryPropertyMap(
+            PropertyMap pm, 
+            TypeMapRegistry registry,
+            ParameterExpression srcParam, 
+            ParameterExpression destParam, 
+            ParameterExpression ctxtParam)
+        {
+            var pmExpression = CreatePropertyMapFunc(pm, registry, srcParam, destParam, ctxtParam);
+
+            if (pmExpression == null)
+                return null;
+
+            var autoMapException = Parameter(typeof(AutoMapperMappingException), "ex");
+            var exception = Parameter(typeof(Exception), "ex");
+
+            var mappingExceptionCtor = typeof(AutoMapperMappingException).GetTypeInfo().DeclaredConstructors.First(ci => ci.GetParameters().Length == 3);
+
+            return TryCatch(Block(typeof(void), pmExpression),
+                MakeCatchBlock(typeof(AutoMapperMappingException), autoMapException, Block(Assign(Property(autoMapException, "PropertyMap"), Constant(pm)), Rethrow()), null),
+                MakeCatchBlock(typeof(Exception), exception, Throw(New(mappingExceptionCtor, ctxtParam, exception, Constant(pm))), null));
+        }
+
+        private static Expression CreateDestinationFunc(
+            TypeMap typeMap,
+            TypeMapRegistry typeMapRegistry,
+            ParameterExpression srcParam,
+            ParameterExpression destParam,
+            ParameterExpression ctxtParam)
+        {
+            var newDestFunc = ToType(CreateNewDestinationFunc(typeMap, typeMapRegistry, srcParam, ctxtParam), typeMap.DestinationType);
+
+            var getDest = typeMap.DestinationTypeToUse.GetTypeInfo().IsValueType
+                ? newDestFunc
+                : Coalesce(destParam, newDestFunc);
+
+            Expression destinationFunc = Assign(destParam, getDest);
+
+            if (typeMap.PreserveReferences)
+            {
+                var dest = Variable(typeof(object), "dest");
+
+                Expression valueBag = Property(ctxtParam, "InstanceCache");
+                var set = Assign(Property(valueBag, "Item", srcParam), dest);
+                var setCache =
+                    IfThen(NotEqual(srcParam, Constant(null)), set);
+
+                destinationFunc = Block(new[] { dest }, Assign(dest, destinationFunc), setCache, dest);
+            }
+            return destinationFunc;
+        }
+
+        private static Expression CreateNewDestinationFunc(
+            TypeMap typeMap,
+            TypeMapRegistry typeMapRegistry, 
+            ParameterExpression srcParam, 
+            ParameterExpression ctxtParam)
+        {
+            if (typeMap.DestinationCtor != null)
+                return typeMap.DestinationCtor.ReplaceParameters(srcParam, ctxtParam);
+
+            if (typeMap.ConstructDestinationUsingServiceLocator)
+                return Call(MakeMemberAccess(ctxtParam, typeof(ResolutionContext).GetProperty("Options")),
+                                typeof(MappingOperationOptions).GetMethod("CreateInstance").MakeGenericMethod(typeMap.DestinationType)
+                                );
+
+            if (typeMap.ConstructorMap?.CanResolve == true)
+                return typeMap.ConstructorMap.BuildExpression(typeMapRegistry, srcParam, ctxtParam);
+
+            if (typeMap.DestinationType.IsInterface())
+            {
+#if PORTABLE
+                Block(typeof(object), Throw(Constant(new PlatformNotSupportedException("Mapping to interfaces through proxies not supported."))), Constant(null));
+#else
+                var ctor = Call(Constant(ObjectCreator.DelegateFactory), typeof(DelegateFactory).GetMethod("CreateCtor", new[] { typeof(Type) }), Call(New(typeof(ProxyGenerator)), typeof(ProxyGenerator).GetMethod("GetProxyType"), Constant(typeMap.DestinationType)));
+                return Invoke(ctor);
+#endif
+            }
+
+            if (typeMap.DestinationType.IsAbstract())
+                return Constant(null);
+
+            if (typeMap.DestinationType.IsGenericTypeDefinition())
+                return Constant(null);
+
+            return DelegateFactory.GenerateConstructorExpression(typeMap.DestinationType);
+        }
+
+        private static Expression CreatePropertyMapFunc(
+            PropertyMap propertyMap, 
+            TypeMapRegistry typeMapRegistry,
+            ParameterExpression srcParam,
+            ParameterExpression destParam,
+            ParameterExpression ctxtParam)
+        {
+            var valueResolverExpr = BuildValueResolverFunc(propertyMap, typeMapRegistry, srcParam, ctxtParam);
+            var destMember = MakeMemberAccess(destParam, propertyMap.DestinationProperty.MemberInfo);
+
+            Expression getter;
+
+            var pi = propertyMap.DestinationProperty.MemberInfo as PropertyInfo;
+            if (pi != null && pi.GetGetMethod(true) == null)
+            {
+                getter = Default(propertyMap.DestinationPropertyType);
+            }
+            else
+            {
+                getter = destMember;
+            }
+
+            var destValueExpr = propertyMap.UseDestinationValue
+                ? getter
+                : Default(propertyMap.TypeMap.DestinationType);
+
+            if (propertyMap.SourceType == null
+                || (propertyMap.SourceType.IsEnumerableType() && propertyMap.SourceType != typeof(string))
+                || typeMapRegistry.GetTypeMap(new TypePair(propertyMap.SourceType, propertyMap.DestinationPropertyType)) != null
+                || ((!EnumMapper.EnumToEnumMapping(new TypePair(propertyMap.SourceType, propertyMap.DestinationPropertyType)) ||
+                  EnumMapper.EnumToNullableTypeMapping(new TypePair(propertyMap.SourceType, propertyMap.DestinationPropertyType))) &&
+                 EnumMapper.EnumToEnumMapping(new TypePair(propertyMap.SourceType, propertyMap.DestinationPropertyType)))
+                || !propertyMap.DestinationPropertyType.IsAssignableFrom(propertyMap.SourceType))
+            {
+                var mapperProp = MakeMemberAccess(ctxtParam, typeof(ResolutionContext).GetProperty("Mapper"));
+                var mapMethod = typeof(IRuntimeMapper)
+                    .GetAllMethods()
+                    .Single(m => m.Name == "Map" && m.IsGenericMethodDefinition)
+                    .MakeGenericMethod(valueResolverExpr.Type, propertyMap.DestinationPropertyType);
+                var second = Call(
+                    mapperProp,
+                    mapMethod,
+                    valueResolverExpr,
+                    destValueExpr,
+                    ctxtParam
+                    );
+                valueResolverExpr = Convert(second, propertyMap.DestinationPropertyType);
+            }
+
+            if (propertyMap.Condition != null)
+            {
+                valueResolverExpr =
+                    Condition(
+                        Invoke(
+                            propertyMap.Condition,
+                            srcParam,
+                            destParam,
+                            ToType(valueResolverExpr, propertyMap.Condition.Parameters[2].Type),
+                            ToType(destValueExpr, propertyMap.Condition.Parameters[2].Type),
+                            ctxtParam
+                            ),
+                        Convert(valueResolverExpr, propertyMap.DestinationPropertyType),
+                        destValueExpr
+                        );
+            }
+
+            Expression mapperExpr;
+            if (propertyMap.DestinationProperty.MemberInfo is FieldInfo)
+            {
+                mapperExpr = propertyMap.SourceType != propertyMap.DestinationPropertyType
+                    ? Assign(destMember, Convert(valueResolverExpr, propertyMap.DestinationPropertyType))
+                    : Assign(getter, valueResolverExpr);
+            }
+            else
+            {
+                var setter = ((PropertyInfo)propertyMap.DestinationProperty.MemberInfo).GetSetMethod(true);
+                if (setter == null)
+                {
+                    mapperExpr = valueResolverExpr;
+                }
+                else
+                {
+                    mapperExpr = Assign(destMember, propertyMap.SourceType != propertyMap.DestinationPropertyType 
+                        ? Convert(valueResolverExpr, propertyMap.DestinationPropertyType) 
+                        : valueResolverExpr);
+                }
+            }
+
+            if (propertyMap.PreCondition != null)
+            {
+                mapperExpr = IfThen(
+                    Invoke(propertyMap.PreCondition, srcParam, ctxtParam),
+                    mapperExpr
+                    );
+            }
+
+            return mapperExpr;
+        }
+
+        private static Expression BuildValueResolverFunc(PropertyMap propertyMap, TypeMapRegistry typeMapRegistry,
+            ParameterExpression srcParam,
+            ParameterExpression ctxtParam)
+        {
+            Expression valueResolverFunc;
+            var valueResolverConfig = propertyMap.ValueResolverConfig;
+            var typeMap = propertyMap.TypeMap;
+
+            if (valueResolverConfig != null)
+            {
+                Expression ctor;
+                Type resolverType;
+                if (valueResolverConfig.Instance != null)
+                {
+                    ctor = Constant(valueResolverConfig.Instance);
+                    resolverType = valueResolverConfig.Instance.GetType();
+                }
+                else
+                {
+                    ctor = Call(MakeMemberAccess(ctxtParam, typeof(ResolutionContext).GetProperty("Options")),
+                                typeof(MappingOperationOptions).GetMethod("CreateInstance").MakeGenericMethod(valueResolverConfig.Type)
+                                );
+                    resolverType = valueResolverConfig.Type;
+                }
+
+                Expression sourceFunc;
+                if (valueResolverConfig.SourceMember != null)
+                {
+                    sourceFunc = valueResolverConfig.SourceMember.ReplaceParameters(srcParam);
+                }
+                else if (valueResolverConfig.SourceMemberName != null)
+                {
+                    sourceFunc = MakeMemberAccess(srcParam,
+                        typeMap.SourceType.GetFieldOrProperty(valueResolverConfig.SourceMemberName));
+                }
+                else
+                {
+                    sourceFunc = srcParam;
+                }
+
+                valueResolverFunc = Convert(Call(ToType(ctor, resolverType), resolverType.GetMethod("Resolve"), sourceFunc, ctxtParam), propertyMap.DestinationPropertyType);
+            }
+            else if (propertyMap.CustomValue != null)
+            {
+                valueResolverFunc = Convert(Constant(propertyMap.CustomValue), propertyMap.DestinationPropertyType);
+            }
+            else if (propertyMap.CustomResolver != null)
+            {
+                valueResolverFunc = TryCatch(Convert(Invoke(Constant(propertyMap.CustomResolver), srcParam, ctxtParam), propertyMap.DestinationPropertyType), Catch(typeof(Exception), Default(propertyMap.DestinationPropertyType)));
+            }
+            else if (propertyMap.CustomExpression != null)
+            {
+                valueResolverFunc = propertyMap.CustomExpression.ReplaceParameters(srcParam).IfNotNull();
+            }
+            else if (propertyMap.SourceMembers.Any()
+                && propertyMap.SourceType != null
+                )
+            {
+                var last = propertyMap.SourceMembers.Last();
+                var pi = last.MemberInfo as PropertyInfo;
+                if (pi != null && pi.GetGetMethod(true) == null)
+                {
+                    valueResolverFunc = Default(last.MemberType);
+                }
+                else
+                {
+                    valueResolverFunc = propertyMap.SourceMembers.Aggregate(
+                        (Expression)srcParam,
+                        (inner, getter) => getter.MemberInfo is MethodInfo
+                            ? getter.MemberInfo.IsStatic()
+                                ? Call(null, (MethodInfo)getter.MemberInfo, inner)
+                                : (Expression)Call(inner, (MethodInfo)getter.MemberInfo)
+                            : MakeMemberAccess(getter.MemberInfo.IsStatic() ? null : inner, getter.MemberInfo)
+                        );
+                    valueResolverFunc = valueResolverFunc.IfNotNull();
+                }
+            }
+            else if (propertyMap.SourceMember != null)
+            {
+                valueResolverFunc = MakeMemberAccess(srcParam, propertyMap.SourceMember);
+            }
+            else
+            {
+                valueResolverFunc = Throw(Constant(new Exception("I done blowed up")));
+            }
+
+            if (propertyMap.DestinationPropertyType == typeof(string) && valueResolverFunc.Type != typeof(string)
+                && typeMapRegistry.GetTypeMap(new TypePair(valueResolverFunc.Type, propertyMap.DestinationPropertyType)) == null)
+            {
+                valueResolverFunc = Call(valueResolverFunc, valueResolverFunc.Type.GetMethod("ToString", new Type[0]));
+            }
+
+            if (propertyMap.NullSubstitute != null)
+            {
+                Expression value = Constant(propertyMap.NullSubstitute);
+                if (propertyMap.NullSubstitute.GetType() != propertyMap.DestinationPropertyType)
+                    value = Convert(value, propertyMap.DestinationPropertyType);
+                valueResolverFunc = MakeBinary(ExpressionType.Coalesce, valueResolverFunc, value);
+            }
+            else if (!typeMap.Profile.AllowNullDestinationValues)
+            {
+                var toCreate = propertyMap.SourceType ?? propertyMap.DestinationPropertyType;
+                if (!toCreate.GetTypeInfo().IsValueType)
+                {
+                    valueResolverFunc = MakeBinary(ExpressionType.Coalesce,
+                        valueResolverFunc,
+                        Convert(Call(
+                            typeof(ObjectCreator).GetMethod("CreateNonNullValue"),
+                            Constant(toCreate)
+                            ), propertyMap.SourceType));
+                }
+            }
+
+            return valueResolverFunc;
+        }
+
+    }
+}
