+++ src/AutoMapper/PropertyMap.cs
@@ -40,12 +40,10 @@ namespace AutoMapper
         private bool _ignored;
         public int? MappingOrder { get; set; }
         public Func<object, ResolutionContext, object> CustomResolver { get; private set; }
-        private bool _sealed;
         public LambdaExpression Condition { get; set; }
         public LambdaExpression PreCondition { get; set; }
         private MemberInfo _sourceMember;
         private LambdaExpression _customExpression;
-        internal LambdaExpression MapExpression { get; private set; }
 
         public PropertyMap(IMemberAccessor destinationProperty, TypeMap typeMap)
         {
@@ -124,25 +122,7 @@ namespace AutoMapper
             CustomValue = CustomValue ?? inheritedMappedProperty.CustomValue;
         }
 
-        internal void Seal(TypeMapRegistry typeMapRegistry)
-        {
-            if (_sealed)
-            {
-                return;
-            }
-
-            if (!CanResolveValue())
-            {
-                MapExpression = null;
-
-                return;
-            }
-
-            MapExpression = this.CreateExpression(typeMapRegistry);
-
-            _sealed = true;
-        }
-
+        // TODO: make this expression based
         public void AssignCustomExpression<TSource, TMember>(Func<TSource, ResolutionContext, TMember> resolverFunc)
         {
             //Expression<Func<TSource, ResolutionContext, TMember>> expr = (s, c) => resolverFunc(s, c);
@@ -359,221 +339,5 @@ namespace AutoMapper
                 : expression;
         }
 
-        public static LambdaExpression CreateExpression(this PropertyMap propertyMap, TypeMapRegistry typeMapRegistry)
-        {
-            var srcParam = Parameter(propertyMap.TypeMap.SourceType, "src");
-            var destParam = Parameter(propertyMap.TypeMap.DestinationType, "dest");
-            var ctxtParam = Parameter(typeof(ResolutionContext), "ctxt");
-            
-            var valueResolverExpr = BuildValueResolverFunc(propertyMap, typeMapRegistry, srcParam, ctxtParam);
-            var destMember = MakeMemberAccess(destParam,propertyMap.DestinationProperty.MemberInfo);
-
-            Expression getter;
-
-            if (propertyMap.DestinationProperty.MemberInfo is PropertyInfo &&
-                ((PropertyInfo)propertyMap.DestinationProperty.MemberInfo).GetGetMethod(true) == null)
-            {
-                getter = Default(propertyMap.DestinationPropertyType);
-            }
-            else
-            {
-                getter = destMember;
-            }
-
-            var destValueExpr = propertyMap.UseDestinationValue
-                ? getter
-                : Default(propertyMap.TypeMap.DestinationType);
-
-            if (propertyMap.SourceType == null
-                || (propertyMap.SourceType.IsEnumerableType() && propertyMap.SourceType != typeof(string))
-                || typeMapRegistry.GetTypeMap(new TypePair(propertyMap.SourceType, propertyMap.DestinationPropertyType)) != null
-                || ((!EnumMapper.EnumToEnumMapping(new TypePair(propertyMap.SourceType, propertyMap.DestinationPropertyType)) ||
-                  EnumMapper.EnumToNullableTypeMapping(new TypePair(propertyMap.SourceType, propertyMap.DestinationPropertyType))) &&
-                 EnumMapper.EnumToEnumMapping(new TypePair(propertyMap.SourceType, propertyMap.DestinationPropertyType)))
-                || !propertyMap.DestinationPropertyType.IsAssignableFrom(propertyMap.SourceType))
-            {
-                var mapperProp = MakeMemberAccess(ctxtParam, typeof(ResolutionContext).GetProperty("Mapper"));
-                var mapMethod = typeof(IRuntimeMapper)
-                    .GetAllMethods()
-                    .Single(m => m.Name == "Map" && m.IsGenericMethodDefinition)
-                    .MakeGenericMethod(valueResolverExpr.Type, propertyMap.DestinationPropertyType);
-                var second = Call(
-                    mapperProp,
-                    mapMethod,
-                    valueResolverExpr,
-                    destValueExpr,
-                    ctxtParam
-                    );
-                valueResolverExpr = Convert(second, propertyMap.DestinationPropertyType);
-            }
-
-            if (propertyMap.Condition != null)
-            {
-                valueResolverExpr =
-                    Condition(
-                        Invoke(
-                            propertyMap.Condition,
-                            srcParam,
-                            destParam,
-                            ToType(valueResolverExpr, propertyMap.Condition.Parameters[2].Type),
-                            ToType(destValueExpr, propertyMap.Condition.Parameters[2].Type),
-                            ctxtParam
-                            ),
-                        Convert(valueResolverExpr, propertyMap.DestinationPropertyType),
-                        destValueExpr
-                        );
-            }
-            
-            Expression mapperExpr;
-            if (propertyMap.DestinationProperty.MemberInfo is FieldInfo)
-            {
-                {
-                    if (propertyMap.SourceType != propertyMap.DestinationPropertyType)
-                        mapperExpr = Assign(destMember, Convert(valueResolverExpr, propertyMap.DestinationPropertyType));
-                    else
-                        mapperExpr = Assign(getter,valueResolverExpr);
-                }
-            }
-            else
-            {
-                var setter = ((PropertyInfo)propertyMap.DestinationProperty.MemberInfo).GetSetMethod(true);
-                if (setter == null)
-                {
-                    mapperExpr = valueResolverExpr;
-                }
-                else
-                {
-                    if (propertyMap.SourceType != propertyMap.DestinationPropertyType)
-                        mapperExpr = Assign(destMember, Convert(valueResolverExpr, propertyMap.DestinationPropertyType));
-                    else
-                        mapperExpr = Assign(destMember, valueResolverExpr);
-                }
-            }
-
-            if (propertyMap.PreCondition != null)
-            {
-                mapperExpr = IfThen(
-                    Invoke(propertyMap.PreCondition, srcParam, ctxtParam),
-                    mapperExpr
-                    );
-            }
-
-            return Lambda(mapperExpr, srcParam, destParam, ctxtParam);
-        }
-
-        private static Expression BuildValueResolverFunc(PropertyMap propertyMap, TypeMapRegistry typeMapRegistry,
-            ParameterExpression srcParam,
-            ParameterExpression ctxtParam)
-        {
-            Expression valueResolverFunc;
-            var valueResolverConfig = propertyMap.ValueResolverConfig;
-            var typeMap = propertyMap.TypeMap;
-
-            if (valueResolverConfig != null)
-            {
-                Expression ctor;
-                Type resolverType;
-                if (valueResolverConfig.Instance != null)
-                {
-                    ctor = Constant(valueResolverConfig.Instance);
-                    resolverType = valueResolverConfig.Instance.GetType();
-                }
-                else
-                {
-                    ctor = Call(MakeMemberAccess(ctxtParam, typeof(ResolutionContext).GetProperty("Options")), 
-                                typeof(MappingOperationOptions).GetMethod("CreateInstance").MakeGenericMethod(valueResolverConfig.Type)
-                                );
-                    resolverType = valueResolverConfig.Type;
-                }
-
-                Expression sourceFunc;
-                if (valueResolverConfig.SourceMember != null)
-                {
-                    sourceFunc = valueResolverConfig.SourceMember.ReplaceParameters(srcParam);
-                }
-                else if (valueResolverConfig.SourceMemberName != null)
-                {
-                    sourceFunc = MakeMemberAccess(srcParam,
-                        typeMap.SourceType.GetFieldOrProperty(valueResolverConfig.SourceMemberName));
-                }
-                else
-                {
-                    sourceFunc = srcParam;
-                }
-                
-                valueResolverFunc = Convert(Call(ToType(ctor, resolverType), resolverType.GetMethod("Resolve"), sourceFunc, ctxtParam), propertyMap.DestinationPropertyType);
-            }
-            else if (propertyMap.CustomValue != null)
-            {
-                valueResolverFunc = Convert(Constant(propertyMap.CustomValue), propertyMap.DestinationPropertyType);
-            }
-            else if (propertyMap.CustomResolver != null)
-            {
-                valueResolverFunc = TryCatch(Convert(Invoke(Constant(propertyMap.CustomResolver), srcParam, ctxtParam), propertyMap.DestinationPropertyType), Catch(typeof(Exception), Default(propertyMap.DestinationPropertyType)));
-            }
-            else if (propertyMap.CustomExpression != null)
-            {
-                valueResolverFunc = propertyMap.CustomExpression.ReplaceParameters(srcParam).IfNotNull();
-            }
-            else if (propertyMap.SourceMembers.Any()
-                && propertyMap.SourceType != null
-                )
-            {
-                var last = propertyMap.SourceMembers.Last();
-                if (last.MemberInfo is PropertyInfo && ((PropertyInfo)last.MemberInfo).GetGetMethod(true) == null)
-                {
-                    valueResolverFunc = Default(last.MemberType);
-                }
-                else
-                {
-                    valueResolverFunc = propertyMap.SourceMembers.Aggregate(
-                        (Expression)srcParam,
-                        (inner, getter) => getter.MemberInfo is MethodInfo
-                            ? getter.MemberInfo.IsStatic()
-                                ? Call(null, (MethodInfo)getter.MemberInfo, inner)
-                                : (Expression)Call(inner, (MethodInfo)getter.MemberInfo)
-                            : MakeMemberAccess(getter.MemberInfo.IsStatic() ? null : inner, getter.MemberInfo)
-                        );
-                    valueResolverFunc = valueResolverFunc.IfNotNull();
-                }
-            }
-            else if (propertyMap.SourceMember != null)
-            {
-                valueResolverFunc = MakeMemberAccess(srcParam, propertyMap.SourceMember);
-            }
-            else
-            {
-                valueResolverFunc = Throw(Constant(new Exception("I done blowed up")));
-            }
-
-            if (propertyMap.DestinationPropertyType == typeof(string) && valueResolverFunc.Type != typeof(string)
-                && typeMapRegistry.GetTypeMap(new TypePair(valueResolverFunc.Type, propertyMap.DestinationPropertyType)) == null)
-            {
-                valueResolverFunc = Call(valueResolverFunc, valueResolverFunc.Type.GetMethod("ToString", new Type[0]));
-            }
-
-            if (propertyMap.NullSubstitute != null)
-            {
-                Expression value = Constant(propertyMap.NullSubstitute);
-                if (propertyMap.NullSubstitute.GetType() != propertyMap.DestinationPropertyType)
-                    value = Convert(value, propertyMap.DestinationPropertyType);
-                valueResolverFunc = MakeBinary(ExpressionType.Coalesce, valueResolverFunc, value);
-            }
-            else if (!typeMap.Profile.AllowNullDestinationValues)
-            {
-                var toCreate = propertyMap.SourceType ?? propertyMap.DestinationPropertyType;
-                if (!toCreate.GetTypeInfo().IsValueType)
-                {
-                    valueResolverFunc = MakeBinary(ExpressionType.Coalesce,
-                        valueResolverFunc,
-                        Convert(Call(
-                            typeof(ObjectCreator).GetMethod("CreateNonNullValue"),
-                            Constant(toCreate)
-                            ), propertyMap.SourceType));
-                }
-            }
-            
-            return valueResolverFunc;
-        }
     }
 }
